requirements.txt
# === Core Simulation Framework ===
metadrive-simulator==0.4.3
gymnasium==0.29.1
numpy==1.26.4
pygame==2.5.2
opencv-python==4.9.0.80
matplotlib==3.8.4

# === Communication & Security ===
cryptography==41.0.7
pycryptodome==3.20.0
psutil==5.9.8

# === Utilities & Config ===
tqdm==4.66.4
PyYAML==6.0.2
cloudpickle==3.0.0
imageio==2.34.1
setuptools==69.5.1
wheel==0.45.1
cmake==3.29.6
ninja==1.11.1
